Turn on support for memcached
The tiddlywebplugins.caching store must be in use for this
module to work. The theory is that if you are going to be going
to the trouble of using this cache, you'll also want the caching store
and thus we can use the same connection to the memcached, because
fewer connections is better.